pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iips"
version = "1.0.0"
description = "Intelligent Invoice Processing System - Multi-agent invoice pipeline"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dateutil>=2.8",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    "pytesseract>=0.3",
    "rapidfuzz>=3.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
iips = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
